Refactor Sound Manager Part 2
now it actually works in the web player
to add new sounds: place footstep sounds in the
Resources/Music/Footsteps folder. MUST use the naming convention tag.#
for your sounds. if this is not observed, then the name will not be
parsed correctly. can look at current footstep sound names for examples